# api/openapi.yaml
openapi: 3.1.0
info:
  title: Relay Station Control API
  version: 0.2.0
  description: |
    Internal API for the single-environment Relay Station Control.

    Except for operations that explicitly set `security: []`, requests require an
    enabled `super_admin` session that satisfies the environment MFA policy. Unsafe
    methods additionally require the session-bound `X-CSRF-Token` header.
security:
  - SessionCookie: []
...
paths:
...
  /api/relay-bindings/nodes/{instance_id}:
    get:
      operationId: getNodeRelayBinding
      summary: Read current Gateway Account binding and resolution for one Relay Node
      description: |
        Returns query-derived resolution and redacted context for the specified Relay Node.
        Never contacts Relay Node or Gateway, and never returns credentials or secrets.
      tags: [relay-bindings]
      parameters:
        - $ref: "#/components/parameters/NodeInstanceId"
      responses:
        "200":
          description: Current Node-centric binding view.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeRelayBindingResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/relay-bindings/gateways/{instance_id}:
    get:
      operationId: getGatewayAccountRelayBindings
      summary: Read Gateway Account-centric binding views for one Gateway Instance
      description: |
        Returns the complete current Directory snapshot accounts and their bound Nodes.
        Never returns credentials or secrets.
      tags: [relay-bindings]
      parameters:
        - $ref: "#/components/parameters/GatewayInstanceId"
      responses:
        "200":
          description: Gateway Account-centric binding envelope.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayAccountRelayBindingsResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/relay-bindings/unresolved:
    get:
      operationId: listUnresolvedRelayBindings
      summary: List all active bindings whose target Account has disappeared from the current fresh Directory snapshot
      description: |
        Returns all unresolved binding relations across registered Gateways.
      tags: [relay-bindings]
      responses:
        "200":
          description: List of unresolved binding relations.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnresolvedRelayBindingsResponse"
        "401":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/relay-bindings/bind:
    post:
      operationId: bindRelayNode
      summary: Bind an unbound Relay Node to a fresh Gateway Account
      description: |
        Requires super_admin session and CSRF proof. Validates Directory freshness (<=540s)
        and account presence within a single transactional boundary.
      tags: [relay-bindings]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BindRelayNodeRequest"
      responses:
        "200":
          description: Relay Node successfully bound to Gateway Account.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelayBindingMutationResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/relay-bindings/rebind:
    post:
      operationId: rebindRelayNode
      summary: Atomically rebind a currently bound Relay Node to another fresh Gateway Account
      description: |
        Requires super_admin session and CSRF proof. Closes the old interval and creates a new
        interval in a single transaction with identical timestamp.
      tags: [relay-bindings]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RebindRelayNodeRequest"
      responses:
        "200":
          description: Relay Node successfully rebound to new Gateway Account.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelayBindingMutationResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/relay-bindings/unbind:
    post:
      operationId: unbindRelayNode
      summary: Unbind a currently bound Relay Node
      description: |
        Requires super_admin session and CSRF proof. Closes open interval with administrator_unbind.
        Idempotent if already unbound.
      tags: [relay-bindings]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnbindRelayNodeRequest"
      responses:
        "200":
          description: Relay Node unbind mutation completed.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelayBindingMutationResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-relay_control_session
      description: |
        Opaque administrator session cookie. In production it is set with Secure,
        HttpOnly, SameSite=Strict, Path=/, no Domain, a 30-minute idle timeout, and a
        12-hour absolute timeout. It is never a bearer token for service identities.
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Random proof bound to the current administrator session.
      schema:
        type: string
        minLength: 32
        writeOnly: true
...
    NodeInstanceId:
      name: instance_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GatewayInstanceId:
      name: instance_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
...
    NoStore:
      description: Sensitive authentication and administrator responses are never cacheable.
      required: true
      schema:
        type: string
        enum:
          - no-store
...
    RequestId:
      description: Stable request correlation identifier safe to show to an administrator.
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
...
  responses:
    Error:
      description: Request rejected using the uniform error envelope.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        Retry-After:
          description: Present for a rate-limited request, in seconds.
          schema:
            type: integer
            minimum: 1
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - authentication_failed
        - mfa_required
        - challenge_expired
        - rate_limited
        - unauthorized
        - forbidden
        - csrf_invalid
        - validation_failed
        - conflict
        - bootstrap_unavailable
        - reauthentication_required
        - last_administrator_protected
        - administrator_self_disable_forbidden
        - not_found
        - internal_error
        - temporarily_unavailable
        - directory_unavailable
        - directory_stale
        - account_not_found
        - node_conflict
        - account_conflict
        - already_unbound
        - no_current_binding
        - asset_not_found
        - asset_retired
        - current_gateway_exists
        - duplicate_identity
        - stale_revision
        - revision_exhausted
        - command_conflict
        - cursor_stale
        - invalid_endpoint
        - secret_configuration_invalid
        - probe_timeout
        - probe_failed
        - service_unavailable
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          minLength: 1
          maxLength: 256
        request_id:
          type: string
          minLength: 8
          maxLength: 128
          pattern: "^[A-Za-z0-9._:-]+$"
        retry_after_seconds:
          type: integer
          minimum: 1
          maximum: 86400
    RelayBindingResolution:
      type: string
      enum: [unbound, resolved, unresolved, unknown]
    RelayBindingFreshness:
      type: string
      enum: [fresh, stale, unavailable]
    RelayBindingContextSource:
      type: string
      enum: [current, last_known, none]
    GatewayAccountId:
      type: string
      minLength: 1
      maxLength: 19
      pattern: "^[1-9][0-9]{0,18}$"
      description: Positive decimal int64 identity, maximum 9223372036854775807. Server rejects overflow. BREAKING correction from numeric JSON; Gateway Directory source v1 remains numeric.
      example: "9007199254740993"
    GatewayAccountContext:
      type: object
      additionalProperties: false
      required: [account_id, name, platform, type, status]
      properties:
        account_id:
          $ref: "#/components/schemas/GatewayAccountId"
        name:
          type: string
        platform:
          type: string
        type:
          type: string
        url:
          type: [string, "null"]
          format: uri
        status:
          type: string
    RelayNodeGatewayAccountBindingDetail:
      type: object
      additionalProperties: false
      required:
        - binding_id
        - relay_node_id
        - gateway_instance_id
        - gateway_account_id
        - evidence_snapshot_id
        - bound_at
        - bound_by
        - bind_reason
      properties:
        binding_id:
          type: string
          format: uuid
        relay_node_id:
          type: string
          format: uuid
        gateway_instance_id:
          type: string
          format: uuid
        gateway_account_id:
          $ref: "#/components/schemas/GatewayAccountId"
        evidence_snapshot_id:
          type: string
          format: uuid
        bound_at:
          type: string
          format: date-time
        bound_by:
          type: string
          format: uuid
        bind_reason:
          type: string
          enum: [administrator_bind, administrator_rebind]
        ended_at:
          type: [string, "null"]
          format: date-time
        ended_by:
          type: [string, "null"]
          format: uuid
        end_reason:
          type: [string, "null"]
          enum: [administrator_unbind, administrator_rebind]
    NodeRelayBindingResponse:
      type: object
      additionalProperties: false
      required:
        - relay_node_id
        - resolution
        - directory_freshness
        - context_source
        - observed_at
      properties:
        relay_node_id:
          type: string
          format: uuid
        current_binding:
          oneOf:
            - $ref: "#/components/schemas/RelayNodeGatewayAccountBindingDetail"
            - type: "null"
        gateway_instance_id:
          type: [string, "null"]
          format: uuid
        gateway_account_id:
          type: [string, "null"]
          minLength: 1
          maxLength: 19
          pattern: "^[1-9][0-9]{0,18}$"
          description: Nullable GatewayAccountId; positive decimal int64, never a JSON number.
        resolution:
          $ref: "#/components/schemas/RelayBindingResolution"
        directory_freshness:
          $ref: "#/components/schemas/RelayBindingFreshness"
        last_success_observation_at:
          type: [string, "null"]
          format: date-time
        context_source:
          $ref: "#/components/schemas/RelayBindingContextSource"
        account_context:
          oneOf:
            - $ref: "#/components/schemas/GatewayAccountContext"
            - type: "null"
        observed_at:
          type: string
          format: date-time
    GatewayAccountCentricBindingItem:
      type: object
      additionalProperties: false
      required:
        - gateway_account_id
        - account_context
        - resolution
        - context_source
      properties:
        gateway_account_id:
          $ref: "#/components/schemas/GatewayAccountId"
        account_context:
          $ref: "#/components/schemas/GatewayAccountContext"
        bound_relay_node_id:
          type: [string, "null"]
          format: uuid
        current_binding:
          oneOf:
            - $ref: "#/components/schemas/RelayNodeGatewayAccountBindingDetail"
            - type: "null"
        resolution:
          $ref: "#/components/schemas/RelayBindingResolution"
        context_source:
          $ref: "#/components/schemas/RelayBindingContextSource"
    GatewayAccountRelayBindingsResponse:
      type: object
      additionalProperties: false
      required:
        - gateway_instance_id
        - directory_freshness
        - observed_at
        - accounts
      properties:
        gateway_instance_id:
          type: string
          format: uuid
        directory_freshness:
          $ref: "#/components/schemas/RelayBindingFreshness"
        current_snapshot_id:
          type: [string, "null"]
          format: uuid
        last_success_observation_at:
          type: [string, "null"]
          format: date-time
        observed_at:
          type: string
          format: date-time
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/GatewayAccountCentricBindingItem"
    UnresolvedRelayBindingItem:
      type: object
      additionalProperties: false
      required:
        - current_binding
        - resolution
        - directory_freshness
        - context_source
        - observed_at
      properties:
        current_binding:
          $ref: "#/components/schemas/RelayNodeGatewayAccountBindingDetail"
        resolution:
          $ref: "#/components/schemas/RelayBindingResolution"
        directory_freshness:
          $ref: "#/components/schemas/RelayBindingFreshness"
        last_success_observation_at:
          type: [string, "null"]
          format: date-time
        context_source:
          $ref: "#/components/schemas/RelayBindingContextSource"
        last_known_account_context:
          oneOf:
            - $ref: "#/components/schemas/GatewayAccountContext"
            - type: "null"
        observed_at:
          type: string
          format: date-time
    UnresolvedRelayBindingsResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UnresolvedRelayBindingItem"
    BindRelayNodeRequest:
      type: object
      additionalProperties: false
      required:
        - relay_node_id
        - gateway_instance_id
        - gateway_account_id
      properties:
        relay_node_id:
          type: string
          format: uuid
        gateway_instance_id:
          type: string
          format: uuid
        gateway_account_id:
          $ref: "#/components/schemas/GatewayAccountId"
    RebindRelayNodeRequest:
      type: object
      additionalProperties: false
      required:
        - relay_node_id
        - new_gateway_instance_id
        - new_gateway_account_id
      properties:
        relay_node_id:
          type: string
          format: uuid
        new_gateway_instance_id:
          type: string
          format: uuid
        new_gateway_account_id:
          $ref: "#/components/schemas/GatewayAccountId"
    UnbindRelayNodeRequest:
      type: object
      additionalProperties: false
      required:
        - relay_node_id
      properties:
        relay_node_id:
          type: string
          format: uuid
    RelayBindingMutationResponse:
      type: object
      additionalProperties: false
      required:
        - outcome
        - operation_at
      properties:
        outcome:
          type: string
          enum: [success, already_unbound]
        binding:
          oneOf:
            - $ref: "#/components/schemas/RelayNodeGatewayAccountBindingDetail"
            - type: "null"
        previous_binding:
          oneOf:
            - $ref: "#/components/schemas/RelayNodeGatewayAccountBindingDetail"
            - type: "null"
        operation_at:
          type: string
          format: date-time
...